CI: make the libcpp build use C++23
Needs a small cast in the qt5 demo viewer to make the addition of two
different enums "ok"

It will help us make sure we don't do things like std::string s = nullptr; (since that is not possible on C++23)

And also helps us future proof in case something else comes up

+++ qt5/demos/viewer.cpp
@@ -54,14 +54,14 @@ PdfViewer::PdfViewer(QWidget *parent) : QMainWindow(parent), m_currentPage(0), m
     // setup the menus
     QMenu *fileMenu = menuBar()->addMenu(tr("&File"));
     m_fileOpenAct = fileMenu->addAction(tr("&Open"), this, &PdfViewer::slotOpenFile);
-    m_fileOpenAct->setShortcut(Qt::CTRL + Qt::Key_O);
+    m_fileOpenAct->setShortcut(static_cast<int>(Qt::CTRL) + Qt::Key_O);
     fileMenu->addSeparator();
     m_fileSaveCopyAct = fileMenu->addAction(tr("&Save a Copy..."), this, &PdfViewer::slotSaveCopy);
     m_fileSaveCopyAct->setShortcut(Qt::CTRL + Qt::SHIFT + Qt::Key_S);
     m_fileSaveCopyAct->setEnabled(false);
     fileMenu->addSeparator();
     QAction *act = fileMenu->addAction(tr("&Quit"), qApp, &QApplication::closeAllWindows);
-    act->setShortcut(Qt::CTRL + Qt::Key_Q);
+    act->setShortcut(static_cast<int>(Qt::CTRL) + Qt::Key_Q);
 
     QMenu *viewMenu = menuBar()->addMenu(tr("&View"));
